refactor: group

+++ src/routes/group/group.service.ts
@@ -1,11 +1,13 @@
 import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { GroupRepo } from './group.repo';
 import { ConversationService } from '../conversation/conversation.service';
+import { ConversationRepo } from '../conversation/conversation.repo';
 
 @Injectable()
 export class GroupService {
   constructor(
     private readonly groupRepo: GroupRepo,
+    private readonly conversationRepo: ConversationRepo,
     private readonly conversationService: ConversationService
   ) { }
 
@@ -59,7 +61,12 @@ export class GroupService {
         HttpStatus.BAD_REQUEST,
       );
     }
-    
+
+    if (!group.conversation?.id) {
+      throw new HttpException('Cannot add', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    await this.conversationRepo.addParticipant(group.conversation?.id, userIdToAdd);
     return this.groupRepo.addMemberToGroup(groupId, userIdToAdd);
   }
 
@@ -72,6 +79,7 @@ export class GroupService {
     if (!group) {
       throw new HttpException('Group not found', HttpStatus.NOT_FOUND);
     }
+
     const isAdmin = await this.groupRepo.isAdmin(currentUserId, groupId);
     if (!isAdmin) {
       throw new HttpException(
@@ -79,9 +87,16 @@ export class GroupService {
         HttpStatus.FORBIDDEN,
       );
     }
+
     if (userIdToRemove === group.adminId) {
       throw new HttpException('Cannot remove admin', HttpStatus.BAD_REQUEST);
     }
+    
+    if (!group.conversation?.id) {
+      throw new HttpException('Cannot remove', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    await this.conversationRepo.removeParticipant(group.conversation?.id, userIdToRemove);
     return this.groupRepo.removeMemberFromGroup(groupId, userIdToRemove);
   }
